Distance Vector Routing
Distance Vector algorithm was developed to route packets from one node
to the other. Each node is made as router, UDP sockets are programmed
using C for packet routing among all the routers. An interface was
provided to update parameters like distance between the routers,
eliminating a router from the network and adding a router to the
network.